Add route to get qualifications
